+++ src/main/java/back/SportApp/User/UserServiceImpl.java
@@ -23,7 +23,7 @@ public class UserServiceImpl implements UserService {
     }
 
     @Override
-    public User findById(Long id) {
+    public User findById(Integer id) {
         Optional<User> user = userRepository.findById(id);
         if (user.isPresent()) {
             return user.get();
@@ -43,7 +43,7 @@ public class UserServiceImpl implements UserService {
     }
 
     @Override
-    public String deleteById(Long id) {
+    public String deleteById(Integer id) {
         userRepository.deleteById(id);
         return "Compte supprimé avec succès.";
     }
